// KairosFrontend/src/app/services/model/userDTO.ts
/**
 * Kairos Backend API
 * This API provides endpoints for the Kairos Backend application.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserImageDTO } from './userImageDTO';

export interface UserDTO { 
    id: any;
    firstName: any;
    lastName: any;
    username: any;
    email: any;
    phoneNumber?: any;
    role: any;
    provider: any;
    profileImage?: UserImageDTO;
}