// KairosFrontend/src/app/services/model/userUpdateDTO.ts
/**
 * Kairos Backend API
 * This API provides endpoints for the Kairos Backend application.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserUpdateDTO { 
    firstName: any;
    lastName: any;
    phoneNumber?: any;
}